+++ firmware/bootloader/bootloader_stubs.cpp
@@ -34,7 +34,7 @@ void extiCallbackThunk(void* data) {
 }
 
 // EXT is not able to give you the front direction but you could read the pin in the callback.
-void efiExtiEnablePin(const char* msg, brain_pin_e brainPin, uint32_t mode, ExtiCallback cb, void* cb_data) {
+bool efiExtiEnablePin(const char* msg, brain_pin_e brainPin, uint32_t mode, ExtiCallback cb, void* cb_data) {
 	ioportid_t port = getHwPort(msg, brainPin);
 	int index = getHwPin(msg, brainPin);
 
@@ -42,6 +42,7 @@ void efiExtiEnablePin(const char* msg, brain_pin_e brainPin, uint32_t mode, Exti
 	palEnableLineEvent(line, mode);
 	callback = cb;
 	palSetLineCallback(line, extiCallbackThunk, cb_data);
+	return true;
 }
 
 void efiExtiDisablePin(brain_pin_e brainPin) {