+++ firmware/hw_layer/digital_input/digital_input_exti.cpp
@@ -34,24 +34,24 @@ struct ExtiChannel {
 static ExtiChannel channels[16];
 
 // EXT is not able to give you the front direction but you could read the pin in the callback.
-void efiExtiEnablePin(const char* msg, brain_pin_e brainPin, uint32_t mode, ExtiCallback cb, void* cb_data) {
+bool efiExtiEnablePin(const char* msg, brain_pin_e brainPin, uint32_t mode, ExtiCallback cb, void* cb_data) {
 	/* paranoid check, in case of Gpio::Unassigned getHwPort will return NULL
 	 * and we will fail on next check */
 	if (!isBrainPinValid(brainPin)) {
-		return;
+		return false;
 	}
 
-	efiAssertVoid(ObdCode::OBD_PCM_Processor_Fault, msg, "efiExtiEnablePin msg must not be null");
+	efiAssert(ObdCode::OBD_PCM_Processor_Fault, msg, "efiExtiEnablePin msg must not be null", false);
 
 	ioportid_t port = getHwPort(msg, brainPin);
 	if (port == NULL) {
-		return;
+		return false;
 	}
 
 	bool wasUsed = brain_pin_markUsed(brainPin, msg);
 	if (wasUsed) {
 		// error condition we shall bail
-		return;
+		return false;
 	}
 
 	int index = getHwPin(msg, brainPin);
@@ -67,7 +67,8 @@ void efiExtiEnablePin(const char* msg, brain_pin_e brainPin, uint32_t mode, Exti
 				hwPortname(brainPin),
 				index,
 				channel.Name);
-		return;
+		brain_pin_markUnused(brainPin);
+		return false;
 	}
 
 	channel.Callback = cb;
@@ -76,6 +77,7 @@ void efiExtiEnablePin(const char* msg, brain_pin_e brainPin, uint32_t mode, Exti
 
 	ioline_t line = PAL_LINE(port, index);
 	palEnableLineEvent(line, mode);
+	return true;
 }
 
 void efiExtiDisablePin(brain_pin_e brainPin) {
@@ -260,7 +262,9 @@ void efiExtiInit() {
 	firmwareError("exti not supported");
 }
 
-void efiExtiEnablePin(const char*, brain_pin_e, uint32_t, ExtiCallback, void*) {}
+bool efiExtiEnablePin(const char*, brain_pin_e, uint32_t, ExtiCallback, void*) {
+	return false;
+}
 void efiExtiDisablePin(brain_pin_e) {}
 
 uint8_t getExtiOverflowCounter() {