+++ firmware/init/sensor/init_sensors.cpp
@@ -63,6 +63,7 @@ void stopSensors() {
 	deinitThermistors();
 	deinitLambda();
 	deInitFlexSensor();
+	deinitHellaOpsT();
 	deinitAuxSensors();
 	deInitVehicleSpeedSensor();
 	deinitTurbochargerSpeedSensor();
@@ -78,6 +79,7 @@ void reconfigureSensors() {
 	initThermistors();
 	initLambda();
 	initFlexSensor();
+	initHellaOpsT();
 	initAuxSensors();
 	initVehicleSpeedSensor();
 	initTurbochargerSpeedSensor();